Be/requirements.txt
alembic==1.15.2
aniso8601==10.0.0
attrs==25.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.24.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.1.3
